Container

Container is the base class for World and Population. It has the basic functionalities to store and access references of other R6::R6Class objects.

Format

R6::R6Class object inheriting from dymiumCore::ContainerGeneric

Construction

x &lt;- Container$new()

NULL

Public Fields

Cont
A named list that contains references of other R6::R6Class objects.

Public Methods

add(x, name)
(R6::R6Class|Supported Transition Models)
Add an R6::R6Class object or a object of the classes in dymiumCore::SupportedTransitionModels to self$Cont. Name must be given when x is an object of the classes in dymiumCore::SupportedTransitionModels.

check_pos(x)
(integer(1)) -> (logical(1))
Check if the position in x exists in self$Cont.

get(x)
(character(1) | integer(1) | an R6::R6Class generator) -> (an R6::R6Class object)
Get a reference to x which can be a R6::R6Class generator or string or integer.

n()
() -> (integer(1))
Returns the length of the references in self$Cont.

names()
() -> (character())
Returns the names of the references in self$Cont.

print()

remove(x)
(character(1) | integer(1) | a R6::R6Class generator)
Remove the reference of x from self$Cont by the class of the given R6::R6Class generator or string or integer.

unpack(target)
(dymiumCore::Container)
Add all the references inside self$Cont to the target container. This is useful when you have containers within a container so that all references can be access via the top container. eg: A World container that contains a Population container can directly access Individual and Household objects.

summary()
Abstract method.