Model class

A container for the supported model objects (see SupportedTransitionModels). When a model object is stored inside Model it can be assess using reference semantics. This is particularly useful when you want to store model objects inside World. By doing so, you can assess those models from World as it is flowing down a microsimulation pipeline just like Entities.

# S3 method for Model
summary(object, ...)

Arguments

object
a Model object

...
dots

Format

R6::R6Class object inheriting Generic.

Construction

x &lt;- Model$new(x)

x :: (caret::train | data.table::data.table | named list)
A model object that compatible.

Active field (read-only)

model
The stored model object in its original form.

Public fields

preprocessing_fn
Default as NULL, this is to store a preprocessing function which will be used to evaluate the entity data in Trans prior to simulating the transition. A situation where this is useful could be when you want to limit the use of a Model object to the specific group of agents (e.g: age between x and y) that was used to estimate the model.

Public Methods

get()
() -> (caret::train | data.table::data.table | named list)
Get a model object.

set(x)
(caret::train | data.table::data.table | named list)
Get a model object.

modify()
An abstract method.

simulate()
An abstract method.

print()

Examples

simple_prob_model <- Model$new(x = list(yes = 0.95, no = 0.05))

simple_glm_model <- Model$new(x = stats::glm(factor(sex) ~ age,
                              data = toy_individuals, family = "binomial"))

# return the original model object
simple_prob_model$model

#> $yes
#> [1] 0.95
#> 
#> $no
#> [1] 0.05
#> 

# add to world
world <- World$new()

world$add(simple_prob_model, name = "simple_prob_model")
world$add(simple_glm_model, name = "simple_glm_model")

# to access
world$get("simple_prob_model")

#> $yes
#> [1] 0.95
#> 
#> $no
#> [1] 0.05
#> 

world$get("simple_glm_model")

#> 
#> Call:  stats::glm(formula = factor(sex) ~ age, family = "binomial", 
#>     data = toy_individuals)
#> 
#> Coefficients:
#> (Intercept)          age  
#>    0.099575    -0.001323  
#> 
#> Degrees of Freedom: 372 Total (i.e. Null);  371 Residual
#> Null Deviance:	    516.9 
#> Residual Deviance: 516.8 	AIC: 520.8

# or alternatively you can use `get_model` which makes sure that it only looks
# through the named list of stored Model objects inside [World].
world$get_model("simple_prob_model")

#> $yes
#> [1] 0.95
#> 
#> $no
#> [1] 0.05
#> 

world$get_model("simple_glm_model")

#> 
#> Call:  stats::glm(formula = factor(sex) ~ age, family = "binomial", 
#>     data = toy_individuals)
#> 
#> Coefficients:
#> (Intercept)          age  
#>    0.099575    -0.001323  
#> 
#> Degrees of Freedom: 372 Total (i.e. Null);  371 Residual
#> Null Deviance:	    516.9 
#> Residual Deviance: 516.8 	AIC: 520.8