Target

Target is to be used within TransitionClassification or supply to event functions. If the target is dynamic then its get will return its target value at the current time or its closest time to the current time.

Format

R6::R6Class object inheriting Generic.

Construction

Target$new(x)

x :: any object that passes check_target()
A target object or NULL.

Active Field (read-only)

data:: a target object
A target object.

dynamic:: logical(1)
A logical flag which indicates whether the target object is dynamic or not.

Public Methods

get(time = .get_sim_time())
(integer(1)) -> a named list()
Get an alignment target as a named list. Note that, all the elements in the list will be scaled by multiplying with the value stored in getOption("dymium.simulation_scale") and all the values in the output are rounded to their nearest integers.

Examples

# static target
TrgtStatic <- Target$new(x = list(yes = 10))
TrgtStatic$data

#> $yes
#> [1] 10
#> 

TrgtStatic$dynamic

#> [1] FALSE

TrgtStatic$get()

#> $yes
#> [1] 10
#> 

# dynamic target
target_dynamic <- data.frame(time = 1:10, yes = 1:10)
TrgtDynamic <- Target$new(x = target_dynamic)
TrgtDynamic$data

#>     time yes
#>  1:    1   1
#>  2:    2   2
#>  3:    3   3
#>  4:    4   4
#>  5:    5   5
#>  6:    6   6
#>  7:    7   7
#>  8:    8   8
#>  9:    9   9
#> 10:   10  10

TrgtDynamic$dynamic

#> [1] TRUE

# if the `time` argument in `get()` is not specified then it will rely on
# the time step from the simulation clock from `.get_sim_time()`.
TrgtDynamic$get()

#> $yes
#> [1] 1
#> 

TrgtDynamic$get(1)

#> $yes
#> [1] 1
#> 

TrgtDynamic$get(10)

#> $yes
#> [1] 10
#>