Transition for regression model

This is particularly useful for updating agents' attributes that can be modelled using a regression model.

By calling the constructor method of TransitionClassification this will initiate the following steps:

initialise(x, model, target, targeted_agents) ->

filter(.data): filter agents to apply the transition to.

mutate(.data): add variables to the data of the filtered agents.

simulate(): simulate the transition outcome using the probabilistic model

postprocess(.sim_result): post-processing the simulation result.

Note that, the order of filter and mutate can be swap by overwriting the mutate_first public field to TRUE. This may be useful in cases where agent selection for the transition depends on one or more derived variables.

To get the simulation result use $get_result().

Format

R6::R6Class object inheriting from Trans

Note

TransitionRegression only returns a raw output of the simulation result. In a future update, there will be an option which allows the prediction result to be drawn from a distribution to add randomness to the result.

Construction

TransitionRegression$new(x, model, targeted_agents = NULL)

x :: R6::R6Class
A Agent class inheritance object.

model :: any object in SupportedTransitionModels
A model object to be used to simulate transition.

targeted_agent :: integer()
(Default as NULL) A integer vectors that contains ids of agents in x to undergo the event.

Fields

mutate_first:: logical(1)
Default as FALSE, this flag is used to indicate whether the attribute data from the Agent in x should be mutated ($mutate(.data)) before filtered ($filter(.data)).

Methods

filter(.data)
(data.table::data.table()) -> [data.table::data.table()]
(By default, first of the preprocessing steps)
By default this method returns the input .data. This method can be overwrite to give the user the flexibility to 'filter' the data prior to making prediction by the given model. Filtering for eligible agents for this transition can be done in this step.

mutate(.data)
(data.table::data.table()) -> [data.table::data.table()]
(By default, second of the preprocessing steps)
By default this method returns the input .data. This method can be overwrite to give the user the flexibility to 'mutate' the data prior to making prediction by the given model. Adding derived variables and historical life course of the agents can be done in this step.

update_agents(attr)
(character(1))
Update the attribute data of the agents that undergo the transition event.

get_result(ids)
(integer()) -> data.table::data.table
Returns the simulation result in a data.table::data.table format with two columns id and response.

get_nrow_result()
Returns the number of rows in the simulation result.

get_decision_maker_ids(response_filter = NULL)
(character()) -> (integer())
Returns ids of the agents that have their response equal to response_filter.

See also

Examples

# load toy data
create_toy_population()

#> [04:20:40] WARN  dymiumCore .subset2(public_bind_env, "initialize"): Creating a `hhsize` column in `hh_data` as it is not provided.

Ind <- pop$get("Individual")

# fit a OLS regression model
model_lm <- glm(age ~ sex + marital_status,
                data = Ind$get_data(),
                family = "gaussian")
summary(model_lm)

#> 
#> Call:
#> glm(formula = age ~ sex + marital_status, family = "gaussian", 
#>     data = Ind$get_data())
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -30.198   -7.738   -0.903    6.659   43.804  
#> 
#> Coefficients:
#>                              Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)                   50.9026     2.5142  20.246  < 2e-16 ***
#> sexmale                        1.1427     1.2156   0.940    0.348    
#> marital_statusmarried          0.2953     2.6325   0.112    0.911    
#> marital_statusnever married  -24.8497     2.7308  -9.100  < 2e-16 ***
#> marital_statusnot applicable -44.3069     2.9462 -15.039  < 2e-16 ***
#> marital_statusseparated       -7.1407     5.3492  -1.335    0.183    
#> marital_statuswidowed         28.4545     4.4281   6.426 4.09e-10 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for gaussian family taken to be 134.7635)
#> 
#>     Null deviance: 168019  on 372  degrees of freedom
#> Residual deviance:  49323  on 366  degrees of freedom
#> AIC: 2896.5
#> 
#> Number of Fisher Scoring iterations: 2
#> 

TransAge <- TransitionRegression$new(Ind, model = model_lm)
# see the simulation result
TransAge

#> There are 373 Individual agents with 12 unique responses of type double {Min. 6.6 | 1st Qu. 26.05 | Median 51.2 | Mean 38.78 | 3rd Qu. 52.34 | Max. 80.5}

# update the individual agents' 'age' field using their simulated age
TransAge$update_agents(attr = "age")

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 80.499912 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 79.357180 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 79.357180 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 43.761967 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 79.357180 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 79.357180 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 44.904699 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 79.357180 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 43.761967 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 80.499912 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 44.904699 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 80.499912 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 43.761967 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 79.357180 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 79.357180 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 26.052986 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 6.595744 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 50.902643 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.045375 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 44.904699 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 7.738476 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 51.197929 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 52.340661 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')

#> Warning: 27.195717 (type 'double') at RHS position 1 truncated (precision lost) when assigning to type 'integer' (column 3 named 'age')