Inspect Entity

Return and print to console the attribute data of entities within an Entity object or its inheritance object. A related entity object can also be queried at the same time.

inspect(entity, ids, related_entity = NULL, verbose = TRUE)

Arguments

entity
an Entity object or its inheritance object

ids
a integer vactor that contains the ids of the entities to be inspected

related_entity
an Entity object or its inheritance object which must be related to entity but not the same as entity. For example, related_entity can be Household if entity is Individual.

verbose
:: logical(1) Should the data be printed to console.

Value

a named ist of data.tables.

Examples

create_toy_world()

#> [10:41:25] WARN  dymiumCore .subset2(public_bind_env, "initialize"): Creating a `hhsize` column in `hh_data` as it is not provided.

inspect(
  entity = world$get("Individual"),
  ids = c(1:3),
  related_entity = world$get("Household")
)

#> ℹ Attribute data of Individual

#>    pid hid age    sex marital_status partner_id father_id mother_id
#> 1:   1   1  71   male  never married         NA        NA        NA
#> 2:   2   2  31 female  never married         NA        NA        NA
#> 3:   3   2  29 female  never married         NA        NA        NA

#> ℹ Attribute data of Household

#> ℹ Note that, entities are 'members' to related entities.

#>    hid did hhsize
#> 1:   1   1      1
#> 2:   2   2      2
#> 3:   2   2      2