Toy transport network data

Note that this dataset requires the sf package to work properly. This dataset contains street network within some SA2 zones of Melborune CBD, Australia in a sf::sf format. This network is contained within the boundaries of toy_zones.

toy_transport_network

Format

A sf::sf data frame with 11 rows and 14 variables:

id

(integer()) zone id

nb_lanes

(integer())

speed

(double())

capacity

(double())

rd_type

(integer())

fnode

(double())

tnode

(double())

geometry

(list())

length

(double())

See also

Examples

toy_transport_network

#> Simple feature collection with 1349 features and 8 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: 144.9074 ymin: -37.85086 xmax: 144.9916 ymax: -37.77596
#> CRS:            EPSG:4283

#> Warning: `...` is not empty.
#> 
#> We detected these problematic arguments:
#> * `needs_dots`
#> 
#> These dots only exist to allow future extensions and should be empty.
#> Did you misspecify an argument?

#> # A tibble: 1,349 x 9
#>       id nb_lanes speed capacity rd_type fnode tnode                  geometry
#>  * <int>    <int> <dbl>    <dbl> <chr>   <dbl> <dbl>          <LINESTRING [°]>
#>  1  8630        2    60     1500 40      64143 64144 (144.9315 -37.78834, 144…
#>  2  8631        2    60     1500 40      64143 64145 (144.9315 -37.78834, 144…
#>  3  8632        1    50      600 46      64143 64146 (144.9315 -37.78834, 144…
#>  4  9374        1    50      600 46      64475 64476 (144.9845 -37.84834, 144…
#>  5  9375        2    60     1700 190250  64475 64477 (144.9845 -37.84834, 144…
#>  6  9376        2    60     1700 190250  64475 64478 (144.9845 -37.84834, 144…
#>  7 10101        1    50      600 46      64476 64475 (144.9848 -37.84839, 144…
#>  8 10177        4    60     3400 190254  65144 65145 (144.9797 -37.84777, 144…
#>  9 10178        3    60     2550 190254  65144 65146 (144.9797 -37.84777, 144…
#> 10 10191        3    70     2850 190255  65160 65161 (144.9895 -37.82031, 144…
#> # … with 1,339 more rows, and 1 more variable: length <dbl>